Split core model listing module
