+++ packages/core/src/model-listing/types.ts
@@ -1,0 +1,17 @@
+export type ListedModel = {
+  id: string;
+  name?: string;
+  description?: string;
+  type?: string;
+  createdAt?: number;
+  ownedBy?: string;
+  contextLength?: number;
+};
+
+export type ModelList = {
+  data: ListedModel[];
+};
+
+export interface ModelListingClient {
+  listModels(): Promise<ModelList>;
+}
